Migrate extensions to their own namespace
