+++ PLRPC.Root/Types/CurrentVersion.cs
@@ -8,19 +8,3 @@ public enum CurrentVersion
     LittleBigPlanet3 = 2,
     LittleBigPlanetVita = 3,
 }
-
-public static class GameVersionExtensions
-{
-    public static string String(this CurrentVersion? version)
-    {
-        return version switch
-        {
-            CurrentVersion.LittleBigPlanet1 => "Playing LittleBigPlanet 1",
-            CurrentVersion.LittleBigPlanet2 => "Playing LittleBigPlanet 2",
-            CurrentVersion.LittleBigPlanet3 => "Playing LittleBigPlanet 3",
-            CurrentVersion.LittleBigPlanetVita => "Playing LittleBigPlanet PS Vita",
-            CurrentVersion.Offline => "Account Offline",
-            _ => ""
-        };
-    }
-}
